test: Split tests
Goes into #204
