+++ tests/test_labels.py
@@ -21,12 +21,11 @@ from capellambse_context_diagrams.collectors import makers
         ),
     ],
 )
-def test_context_diagrams(
+def test_label_wrapping(
     model: capellambse.MelodyModel, uuid: str, expected_labels: list[list[str]]
-) -> None:
+):
     obj = model.by_uuid(uuid)
 
     labels = makers.make_label(obj.name, max_width=makers.MAX_LABEL_WIDTH)
 
-    actual = [label.text for label in labels]
-    assert actual == expected_labels
+    assert [label.text for label in labels] == expected_labels
